competitor researcher done
