+++ examples/competitor_researcher/.env.example
@@ -1,0 +1,4 @@
+OPENAI_API_KEY=KEY
+MODEL=gpt-4o
+URL="https://mintlify.com/docs/quickstart"
+NOTION_PARENT_PAGE=PARENT_PAGE
