+++ examples/competitor_researcher/README.md
@@ -1,10 +1,11 @@
-# Calendar Agent
+# Competitor Researcher
 > Fork and Clone this repository if needed!
 
 ## Introduction
-This project is an example which uses Composio to seamlessly convert your to-do lists into Google Calendar events. 
-It automatically schedules tasks with specified labels and times, ensuring your calendar is always up-to-date and organized.
-
+This project uses Composio to automate the creation and management of competitor pages in Notion. 
+It scrapes data from competitor websites and generates Notion pages under a specified parent page. 
+If a page with the same name already exists, a unique identifier is added. 
+This ensures that your competitor information in Notion is always up-to-date and well-organized.
 ## Steps to Run
 **Navigate to the Project Directory:**
 Change to the directory where the `setup.sh`, `main.py`, `requirements.txt`, and `README.md` files are located. For example:
@@ -24,9 +25,8 @@ add necessary tools:
 ./setup.sh
 ```
 Now, Fill in the .env file with your secrets.
-
 ### 2. Run the python script
 ```shell
-python cookbook/examples/calendar_agent/main.py
+python cookbook/examples/commit_agent/main.py
 ```
-A new event has been added to our Google calendar!
+Your notion page should automatically be populated with the data.
