+++ examples/competitor_researcher/requirements.txt
@@ -1,0 +1,7 @@
+python-dotenv
+requests
+beautifulsoup4
+composio-crewai
+crewai
+flask
+langchain-openai
